+++ src/main/java/compression/timestamp/TimeStampCompressionModel.java
@@ -7,7 +7,7 @@ import java.util.List;
 import java.util.function.Function;
 
 public abstract class TimeStampCompressionModel extends BaseModel<Long> {
-    public TimeStampCompressionModel(Double errorBound, Integer lengthBound) {
+    public TimeStampCompressionModel(Float errorBound, Integer lengthBound) {
         super(errorBound, lengthBound);
     }
