+++ src/main/java/config/ConfigProperties.java
@@ -66,8 +66,8 @@ public class ConfigProperties extends Properties{
         return Double.parseDouble(getProperty("model.timestamp.errorbound"));
     }
 
-    public double getValueModelErrorBound(){
-        return Double.parseDouble(getProperty("model.value.errorbound"));
+    public float getValueModelErrorBound(){
+        return Float.parseFloat(getProperty("model.value.errorbound"));
     }
 
     public String getJDBConnectionString(){
